video/ai-content-generator/requirements.txt
# AI Content Generator - Python 依赖

# CLI 框架
typer>=0.9.0
click>=8.0.0

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# HTTP 请求
requests>=2.31.0
httpx>=0.25.0

# 异步支持（可选）
aiohttp>=3.9.0

# 图片处理（可选，用于后处理）
Pillow>=10.0.0

# 进度条
tqdm>=4.66.0

# 开发依赖
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
